Namespaces are cool, let's use them more
Specifically we move ``StagedPetitionCreator`` into the ``Staged`` namespace as ``Staged::PetitionCreator``.  Move most of the models that were in ``Staged`` to ``Staged::Petition`` because they're really about staged petitions and we're about to add something else that can be staged.
