--- app/models/staged/petition.rb
@@ -1,5 +1,0 @@
-module Staged
-  class Petition < Staged::Base::Petition
-    include Staged::Validations::PetitionDetails
-  end
-end
