+++ app/models/staged/petition/creator.rb
@@ -1,0 +1,13 @@
+module Staged
+  module Petition
+    class Creator < Staged::Base::Petition
+      include Staged::Petition::HasCreatorSignature
+
+      class CreatorSignature < Staged::Base::CreatorSignature
+        include Staged::Validations::SignerDetails
+        include Staged::Validations::Email
+      end
+    end
+  end
+end
+
