+++ app/models/staged/petition/petition.rb
@@ -1,0 +1,7 @@
+module Staged
+  module Petition
+    class Petition < Staged::Base::Petition
+      include Staged::Validations::PetitionDetails
+    end
+  end
+end
