+++ app/models/staged/petition/replay_email.rb
@@ -1,0 +1,11 @@
+module Staged
+  module Petition
+    class ReplayEmail < Staged::Base::Petition
+      include Staged::Petition::HasCreatorSignature
+
+      class CreatorSignature < Staged::Base::CreatorSignature
+        include Staged::Validations::Email
+      end
+    end
+  end
+end
