+++ app/models/staged/petition/replay_petition.rb
@@ -1,0 +1,10 @@
+module Staged
+  module Petition
+    class ReplayPetition < Staged::Base::Petition
+      include Staged::Petition::HasCreatorSignature
+
+      class CreatorSignature < Staged::Base::CreatorSignature
+      end
+    end
+  end
+end
