--- app/models/staged/replay_petition.rb
@@ -1,8 +1,0 @@
-module Staged
-  class ReplayPetition < Staged::Base::Petition
-    include Staged::HasCreatorSignature
-
-    class CreatorSignature < Staged::Base::CreatorSignature
-    end
-  end
-end
